done. minor little features I'd like to add before im really, really done, but it works beautifully and the interface is nice and the scrollbar is shitty sometimes but it never breaks.

+++ CellularAutomataCustom/Main.cpp
@@ -41,7 +41,7 @@ void main()
 				}
 				break;
 			case sf::Event::MouseWheelMoved:
-				input::MouseWheelDelta = e.mouseWheel.delta;
+				input::MouseWheelDelta = (float)e.mouseWheel.delta;
 				break;
 			}
